ai_music_generator_agent/pyproject.toml
[project]
name = "ai-music-generator-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.3",
    "google-generativeai>=0.8.5",
    "numpy>=2.3.5",
    "python-dotenv>=1.2.1",
    "scipy>=1.16.3",
    "uvicorn>=0.38.0",
]
